fixed error

+++ page/checkout/confirm.php
@@ -17,7 +17,7 @@ $get_cart = get('cart', 'WHERE user_id=' . $user_id);
 foreach ($get_cart as $data_cart) {
     $product_id = $data_cart['product_id'];
     $quantity = $data_cart['quantity'];
-    
+
     $get_product = get('product', 'WHERE product_id=' . $product_id);
     $data_product = mysqli_fetch_assoc($get_product);
 
@@ -29,7 +29,7 @@ foreach ($get_cart as $data_cart) {
     } else {
         $price = $data_product['price'];
     }
-    
+
     insert('transaction_details', [
         'transaction_id' => $last_id,
         'product_id' => $product_id,
@@ -37,18 +37,17 @@ foreach ($get_cart as $data_cart) {
         'quantity' => $quantity
     ]);
 
-    // var_dump($connect->insert_id);
-
     $sold = $data_product['sold'];
     $sold = (int)$sold + (int)$quantity;
 
     $query = 'UPDATE product SET sold=' . $sold . ' WHERE product_id=' . $product_id;
     mysqli_query($connect, $query);
 
-    $query = "DELETE FROM cart WHERE product_id=" . $product_id;
+    $query = "DELETE FROM cart WHERE product_id=" . $product_id . " AND user_id=" . $user_id;
+    mysqli_query($connect, $query);
+
+    $query = "DELETE FROM wishlist WHERE product_id=" . $product_id . " AND user_id=" . $user_id;
     mysqli_query($connect, $query);
 }
-
-var_dump(' outside foreach '.$connect->insert_id);
 
 echo "<script>window.location.href = 'index.php?page=cart_list'</script>";
